# SqlClient improve managed memory usage