The managed Tds implementation is only used on Unix where the native implementation cannot be used, this is for performance reasons. I've investigated and made some changes which improve the performance of the managed mode. There is a still a lot that can be done but I think they'll need to build on this start.

The release build now works so I can get inproc BDN numbers:

|      Method |     Mean |    Error |   StdDev |   Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------ |---------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
| SyncFloat32 | 235.2 ms | 4.871 ms | 12.31 ms | 228.8 ms |   3000.0000 |           - |           - |           593.75 KB |

|      Method |     Mean |    Error |   StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------ |---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
| SyncFloat32 | 229.9 ms | 3.993 ms | 3.540 ms |           - |           - |           - |           281.25 KB |

Which shows that there is a large improvement in memory usage. The numbers are a bit sterile though. Compare profiles from before where there is a gen0 roughly every 100ms.

![before](https://user-images.githubusercontent.com/13322696/50405553-7c24c800-07ae-11e9-8c7c-eafb450be3f4.PNG)

to after:

![after](https://user-images.githubusercontent.com/13322696/50405557-834bd600-07ae-11e9-9804-68878572eed6.PNG)

The changes are straightforward in terms of code changes even though it was tricky working out the right places to make them. Packets were being dropped without disposing them causing their rented arrays to be lost, this meant a lot of churn in array allocations. The mars mux header writing code was allocating a lot of temporary arrays so I replaced that code with a leaner version and stopped re-allocation of the header itself where it could be re-used.

The changes pass standard and manual sql tests.

/cc all the usual people, @saurabh500 @AfsanehR @keeratsingh