@Wraith2 2018-12-25
I identified all the tests that break with the assertion covered in corefx/33930 and marked them with activeissue since there are only 12 of them. I could also have run the manual tests in release mode but that way I wouldn't have seen any other assertions and I think the diagnostics of a failure would have been helpful.