@saurabh500 2018-12-25
All on Windows with switch flipped? Can you list the tests that are disabled? 

What does you benchmark test?