@Wraith2 2018-12-25
Yes, all windows with the switch flipped, My local version lets me set it in both release and managed mode but otherwise it's identical. 

The benchmark is a single connection, single command, sproc insert of a float into a table. [code is here](https://github.com/Wraith2/SqlBench/blob/1e26aee93bc81c008bc51581537028b492b39ad7/Program.cs#L196). The intention is just to test simple small packets passing back and forth.

I've added the list of disabled tests to https://github.com/dotnet/corefx/issues/33930 . The Tvp test is also not unix specific and this is covered in https://github.com/dotnet/corefx/issues/27858