@Wraith2 2018-12-25
Yes. And there are only two mars tests disabled, all the others pass. The hard work of muxing is all untouched I've only changed how the bytes are read/written to the buffer slightly and moved that logic into the smux header class.