@Wraith2 2018-12-25
@dotnet-bot test Linux arm64 Release Build
@dotnet-bot test Packaging All Configurations x64 Debug Build