@Wraith2 2018-12-25
Ah, I see them now in the packaging version. Just BitConverter overload stuff which is easily fixed. Thanks for pointing it out.