@Wraith2 2019-01-03
Using this basis I have a branch which makes further improvements.
![new](https://user-images.githubusercontent.com/13322696/50654958-7d13e180-0f86-11e9-915e-d949ea2ad8a2.jpg)