@AfsanehR 2019-01-10
@dotnet-bot test Windows x64 Debug Build
@dotnet-bot test Windows x86 Release Build